Models: change num_stars to INTEGER

+++ db/Hotel.js
@@ -2,7 +2,7 @@ const tripDB = require( './db' );
 
 const Hotel = tripDB.define('hotel', {
     name: tripDB.Sequelize.STRING,
-    num_stars: tripDB.Sequelize.FLOAT,
+    num_stars: tripDB.Sequelize.INTEGER,
     amenities: tripDB.Sequelize.STRING
 });
